tg_product/bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0